requirements.txt
fastapi==0.115.12
langchain==0.3.25
langchain_openai==0.3.16
neo4j==5.28.0
numpy==2.2.5
pandas==2.2.3
pydantic==2.11.4
python-dotenv==1.1.0
Requests==2.32.3
scikit_learn==1.6.1
streamlit==1.42.0
torch==2.6.0+cu118
transformers==4.49.0
